fix test script
- add local testing
- always upgrade pip

+++ test.sh
@@ -4,8 +4,10 @@
 
 cd $(dirname $0) && set -xe
 
-DOCKERNAME=automon
-
-./build.sh
-
-docker run --rm $DOCKERNAME test "$@"
+if [ "$1" == "--local" ]; then
+  python3 -m pytest automon
+else
+  DOCKERNAME=automon
+  ./build.sh
+  docker run --rm $DOCKERNAME test "$@"
+fi
